Revert "Introduce auth_ident_file setting and enhance cert and peer auth methods with user name maps"
