+++ etc/pgbouncer.ini
@@ -125,9 +125,6 @@ auth_file = /etc/pgbouncer/userlist.txt
 ;; Path to HBA-style auth config
 ;auth_hba_file =
 
-;; Path to Pg-ident-style map file
-; auth_ident_file =
-
 ;; Query to use to fetch password from database.  Result
 ;; must have 2 columns - username and password hash.
 ;auth_query = SELECT usename, passwd FROM pg_shadow WHERE usename=$1
