+++ include/hba.h
@@ -16,61 +16,8 @@
  * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  */
 
-#define NAME_ALL        1
-#define NAME_SAMEUSER   2
+struct HBA;
 
-enum RuleType {
-	RULE_LOCAL,
-	RULE_HOST,
-	RULE_HOSTSSL,
-	RULE_HOSTNOSSL,
-};
-
-struct NameSlot {
-	size_t strlen;
-	char str[];
-};
-struct HBAName {
-	unsigned int flags;
-	struct StrSet *name_set;
-};
-
-struct HBARule {
-	struct List node;
-	enum RuleType rule_type;
-	int rule_method;
-	int rule_af;
-	uint8_t rule_addr[16];
-	uint8_t rule_mask[16];
-	struct HBAName db_name;
-	struct HBAName user_name;
-	struct IdentMap *identmap;
-	int hba_linenr;
-};
-
-struct HBA {
-	struct List rules;
-};
-
-struct Mapping {
-	struct List node;
-	char *system_user_name;
-	char *postgres_user_name;
-	unsigned int name_flags;
-};
-
-struct IdentMap {
-	struct List node;
-	char *map_name;
-	struct List mappings;
-};
-
-struct Ident {
-	struct List maps;
-};
-
-struct Ident *ident_load_map(const char *fn);
-void ident_free(struct Ident *ident);
-struct HBA *hba_load_rules(const char *fn, struct Ident *ident);
+struct HBA *hba_load_rules(const char *fn);
 void hba_free(struct HBA *hba);
-struct HBARule *hba_eval(struct HBA *hba, PgAddr *addr, bool is_tls, const char *dbname, const char *username);
+int hba_eval(struct HBA *hba, PgAddr *addr, bool is_tls, const char *dbname, const char *username);
