+++ src/hba.c
@@ -24,6 +24,41 @@
 #include <usual/socket.h>
 #include <usual/string.h>
 
+enum RuleType {
+	RULE_LOCAL,
+	RULE_HOST,
+	RULE_HOSTSSL,
+	RULE_HOSTNOSSL,
+};
+
+#define NAME_ALL        1
+#define NAME_SAMEUSER   2
+
+struct NameSlot {
+	size_t strlen;
+	char str[];
+};
+
+struct HBAName {
+	unsigned int flags;
+	struct StrSet *name_set;
+};
+
+struct HBARule {
+	struct List node;
+	enum RuleType rule_type;
+	int rule_method;
+	int rule_af;
+	uint8_t rule_addr[16];
+	uint8_t rule_mask[16];
+	struct HBAName db_name;
+	struct HBAName user_name;
+};
+
+struct HBA {
+	struct List rules;
+};
+
 /*
  * StrSet
  */
@@ -310,7 +345,6 @@ static void free_parser(struct TokParser *p)
 }
 
 static bool parse_names(struct HBAName *hname, struct TokParser *p, bool is_db, const char *parent_filename);
-static bool parse_ident_name(const char **ident_name, struct TokParser *tp, bool *is_name_all);
 
 static bool parse_namefile(struct HBAName *hname, const char *fn, bool is_db)
 {
@@ -341,22 +375,6 @@ static bool parse_namefile(struct HBAName *hname, const char *fn, bool is_db)
 	free(ln);
 	fclose(f);
 	return ok;
-}
-
-static bool parse_ident_name(const char **ident_name, struct TokParser *tp, bool *is_name_all)
-{
-	if (eat_kw(tp, "all")) {
-		*is_name_all = true;
-		return true;
-	}
-
-	if (!expect(tp, TOK_IDENT, ident_name)) {
-		if (!expect(tp, TOK_STRING, ident_name)) {
-			return false;
-		}
-	}
-
-	return true;
 }
 
 static bool parse_names(struct HBAName *hname, struct TokParser *tp, bool is_db, const char *parent_filename)
@@ -481,175 +499,7 @@ static bool bad_mask(struct HBARule *rule)
 	return !!res;
 }
 
-static bool match_map(struct HBARule *rule, struct Ident *ident, const char *mapname)
-{
-	struct List *el;
-	struct IdentMap *map;
-
-	if (!ident)
-		return false;
-
-	list_for_each(el, &ident->maps) {
-		map = container_of(el, struct IdentMap, node);
-
-		if (strcmp(map->map_name, mapname) == 0) {
-			rule->identmap = map;
-			return true;
-		}
-	}
-
-	return false;
-}
-
-static bool parse_map_definition(struct HBARule *rule, struct Ident *ident, struct TokParser *tp, int linenr)
-{
-	const char *str;
-	char *val;
-
-	if (!expect(tp, TOK_IDENT, &str))
-		return true;
-
-	val = strchr(str, '=');
-
-	if (val == NULL || strncmp(str, "map=", 4) != 0) {
-		log_warning("hba line %d: Ident map %s is malformed. It is not in map=value format.", linenr, str);
-		return false;
-	}
-
-	val++;
-
-	next_token(tp);
-
-	if (!match_map(rule, ident, val)) {
-		log_warning("hba line %d: Ident map %s is not found in ident config file", linenr, val);
-		return false;
-	}
-
-	return true;
-}
-static void mapping_free(struct Mapping *mapping)
-{
-	free(mapping->system_user_name);
-	free(mapping->postgres_user_name);
-	free(mapping);
-}
-
-static void ident_map_free(struct IdentMap *ident_map)
-{
-	struct List *el, *tmp;
-	struct Mapping *mapping;
-
-	if (!ident_map)
-		return;
-
-	list_for_each_safe(el, &ident_map->mappings, tmp) {
-		mapping = container_of(el, struct Mapping, node);
-		list_del(&mapping->node);
-		mapping_free(mapping);
-	}
-
-	free(ident_map->map_name);
-	free(ident_map);
-}
-
-static bool find_ident_map(struct Ident *ident, const char *mapname, struct IdentMap **ident_map)
-{
-	struct List *el;
-
-	list_for_each(el, &ident->maps) {
-		*ident_map = container_of(el, struct IdentMap, node);
-
-		if (!strcmp((*ident_map)->map_name, mapname))
-			return true;
-	}
-
-	return false;
-}
-
-static bool parse_ident_line(struct Ident *ident, struct TokParser *tp, int linenr)
-{
-	const char *map_name = NULL;
-	char *map_name_copy = NULL;
-	const char *system_user_name = NULL;
-	const char *postgres_user_name = NULL;
-	struct IdentMap *ident_map = NULL;
-	struct Mapping *mapping = NULL;
-
-	bool is_name_all = false;
-
-	if (eat(tp, TOK_EOL)) {
-		return true;
-	}
-
-	mapping = calloc(sizeof *mapping, 1);
-
-	if (!mapping) {
-		log_warning("ident: no mem for parsing mapping");
-		return false;
-	}
-
-	if (!expect(tp, TOK_IDENT, &map_name)) {
-		goto failed;
-	}
-
-	map_name_copy = strdup(map_name);
-
-	next_token(tp);
-
-	if (!expect(tp, TOK_IDENT, &system_user_name)) {
-		if (!expect(tp, TOK_STRING, &system_user_name))
-			goto failed;
-	}
-
-	mapping->system_user_name = strdup(system_user_name);
-
-	next_token(tp);
-
-
-	if (!parse_ident_name(&postgres_user_name, tp, &is_name_all)) {
-		goto failed;
-	}
-
-	if (is_name_all) {
-		mapping->name_flags |= NAME_ALL;
-	} else {
-		mapping->postgres_user_name = strdup(postgres_user_name);
-	}
-
-	next_token(tp);
-
-	if (!eat(tp, TOK_EOL)) {
-		log_warning("ident line %d: unsupported parameters", linenr);
-		goto failed;
-	}
-
-
-	if (find_ident_map(ident, map_name_copy, &ident_map)) {
-		list_append(&ident_map->mappings, &mapping->node);
-		free(map_name_copy);
-	} else {
-		ident_map = calloc(sizeof *ident_map, 1);
-
-		if (!ident_map) {
-			log_warning("ident: no mem for parsing ident_map");
-			return false;
-		}
-
-		ident_map->map_name = map_name_copy;
-		list_init(&ident_map->mappings);
-		list_append(&ident_map->mappings, &mapping->node);
-		list_append(&ident->maps, &ident_map->node);
-	}
-
-	return true;
-
-failed:
-	mapping_free(mapping);
-	ident_map_free(ident_map);
-	return false;
-}
-
-static bool parse_line(struct HBA *hba, struct Ident *ident, struct TokParser *tp, int linenr, const char *parent_filename)
+static bool parse_line(struct HBA *hba, struct TokParser *tp, int linenr, const char *parent_filename)
 {
 	const char *addr = NULL, *mask = NULL;
 	enum RuleType rtype;
@@ -745,16 +595,11 @@ static bool parse_line(struct HBA *hba, struct Ident *ident, struct TokParser *t
 		goto failed;
 	}
 
-	if (!parse_map_definition(rule, ident, tp, linenr)) {
-		goto failed;
-	}
-
 	if (!eat(tp, TOK_EOL)) {
 		log_warning("hba line %d: unsupported parameters", linenr);
 		goto failed;
 	}
 
-	rule->hba_linenr = linenr;
 	list_append(&hba->rules, &rule->node);
 	return true;
 failed:
@@ -762,61 +607,7 @@ failed:
 	return false;
 }
 
-struct Ident *ident_load_map(const char *fn)
-{
-	struct Ident *ident = NULL;
-	FILE *f = NULL;
-	char *ln = NULL;
-	size_t lnbuf = 0;
-	ssize_t len;
-	int linenr;
-
-	struct TokParser tp;
-
-	if (fn == NULL)
-		return NULL;
-
-	init_parser(&tp);
-
-	ident = malloc(sizeof *ident);
-
-	list_init(&ident->maps);
-
-	if (!ident)
-		goto out;
-
-	f = fopen(fn, "r");
-
-	if (!f) {
-		log_error("could not open ident config file %s: %s", fn, strerror(errno));
-		goto out;
-	}
-
-	for (linenr = 1; ; linenr++) {
-		len = getline(&ln, &lnbuf, f);
-		if (len < 0)
-			break;
-
-		parse_from_string(&tp, ln);
-
-		if (!parse_ident_line(ident, &tp, linenr)) {
-			/* Tell the admin where to look for the problem. */
-			log_warning("could not parse ident config line %d", linenr);
-			/* Ignore line, but parse to the end. */
-			continue;
-		}
-	}
-
-out:
-	free_parser(&tp);
-	free(ln);
-
-	if (f)
-		fclose(f);
-	return ident;
-}
-
-struct HBA *hba_load_rules(const char *fn, struct Ident *ident)
+struct HBA *hba_load_rules(const char *fn)
 {
 	struct HBA *hba = NULL;
 	FILE *f = NULL;
@@ -845,7 +636,7 @@ struct HBA *hba_load_rules(const char *fn, struct Ident *ident)
 		if (len < 0)
 			break;
 		parse_from_string(&tp, ln);
-		if (!parse_line(hba, ident, &tp, linenr, fn)) {
+		if (!parse_line(hba, &tp, linenr, fn)) {
 			/* Tell the admin where to look for the problem. */
 			log_warning("could not parse hba config line %d", linenr);
 			/* Ignore line, but parse to the end. */
@@ -858,22 +649,6 @@ out:
 	if (f)
 		fclose(f);
 	return hba;
-}
-
-void ident_free(struct Ident *ident)
-{
-	struct List *el, *tmp;
-	struct IdentMap *map;
-
-	if (!ident)
-		return;
-
-	list_for_each_safe(el, &ident->maps, tmp) {
-		map = container_of(el, struct IdentMap, node);
-		list_del(&map->node);
-		ident_map_free(map);
-	}
-	free(ident);
 }
 
 void hba_free(struct HBA *hba)
@@ -924,7 +699,7 @@ static bool match_inet6(const struct HBARule *rule, PgAddr *addr)
 	       (src[2] & mask[2]) == base[2] && (src[3] & mask[3]) == base[3];
 }
 
-struct HBARule * hba_eval(struct HBA *hba, PgAddr *addr, bool is_tls, const char *dbname, const char *username)
+int hba_eval(struct HBA *hba, PgAddr *addr, bool is_tls, const char *dbname, const char *username)
 {
 	struct List *el;
 	struct HBARule *rule;
@@ -932,7 +707,7 @@ struct HBARule * hba_eval(struct HBA *hba, PgAddr *addr, bool is_tls, const char
 	unsigned int unamelen = strlen(username);
 
 	if (!hba)
-		return NULL;
+		return AUTH_REJECT;
 
 	list_for_each(el, &hba->rules) {
 		rule = container_of(el, struct HBARule, node);
@@ -964,7 +739,7 @@ struct HBARule * hba_eval(struct HBA *hba, PgAddr *addr, bool is_tls, const char
 			continue;
 
 		/* rule matches */
-		return rule;
+		return rule->rule_method;
 	}
-	return NULL;
+	return AUTH_REJECT;
 }
