--- test/pgbouncer_hba.conf
@@ -1,14 +1,0 @@
-# TYPE  DATABASE        USER            ADDRESS                 METHOD
-local   all             postgres                                peer
-# do not let the "postgres" superuser login via a certificate
-hostssl all             postgres        ::/0                    reject
-hostssl all             postgres        0.0.0.0/0               reject
-#
-hostssl all             all             ::/0                    md5
-
-hostssl p0y             all             0.0.0.0/0               cert    map=test2
-hostssl p0x             all             0.0.0.0/0               cert    map=test
-hostssl p0              bouncer         0.0.0.0/0               cert
-
-hostssl all             all		192.168.1.1/0		md5
-
