--- test/pgident.conf
@@ -1,9 +1,0 @@
-# PostgreSQL User Name Maps
-# =========================
-
-# MAPNAME       SYSTEM-USERNAME         PG-USERNAME
-test		pgbouncer.acme.org	someuser
-test		pgbouncer.acme.org	anotheruser
-test2		"bouncer"		all
-test2		pgbouncer.acme.org	"anotheruser"
-
