+++ test/test_auth.py
@@ -1,12 +1,10 @@
-import getpass
 import re
-import subprocess
 import time
 
 import psycopg
 import pytest
 
-from .utils import LONG_PASSWORD, MACOS, PG_SUPPORTS_SCRAM, TLS_SUPPORT, WINDOWS
+from .utils import LONG_PASSWORD, PG_SUPPORTS_SCRAM, WINDOWS
 
 
 @pytest.mark.md5
@@ -578,232 +576,3 @@ async def test_change_server_password_server_lifetime(bouncer, pg):
                 await result3
     finally:
         pg.sql("ALTER USER puser1 PASSWORD 'foo'")
-
-
-@pytest.mark.skipif("MACOS", reason="SSL tests are broken on OSX in CI #1031")
-@pytest.mark.skipif("WINDOWS", reason="Windows does not have SIGHUP")
-@pytest.mark.skipif(not TLS_SUPPORT, reason="pgbouncer is built without TLS support")
-def test_client_hba_cert(bouncer, cert_dir):
-    root = cert_dir / "TestCA1" / "ca.crt"
-    key = cert_dir / "TestCA1" / "sites" / "01-localhost.key"
-    cert = cert_dir / "TestCA1" / "sites" / "01-localhost.crt"
-
-    bouncer.write_ini(f"client_tls_key_file = {key}")
-    bouncer.write_ini(f"client_tls_cert_file = {cert}")
-    bouncer.write_ini(f"client_tls_ca_file = {root}")
-    bouncer.write_ini(f"client_tls_sslmode = require")
-    bouncer.write_ini(f"auth_type = hba")
-    bouncer.write_ini(
-        f"auth_query = SELECT usename, passwd FROM pg_shadow where usename = $1"
-    )
-    bouncer.write_ini(f"auth_user = pswcheck")
-    bouncer.write_ini(f"auth_file = {bouncer.auth_path}")
-    bouncer.write_ini(f"auth_hba_file = pgbouncer_hba.conf")
-    bouncer.write_ini(f"auth_ident_file = pgident.conf")
-
-    bouncer.admin("reload")
-
-    client_key = cert_dir / "TestCA1" / "sites" / "04-pgbouncer.acme.org.key"
-    client_cert = cert_dir / "TestCA1" / "sites" / "04-pgbouncer.acme.org.crt"
-
-    # The client connects to p0x using a client certificate with CN=pgbouncer.acme.org.
-    # hba_eval returns the followign line:
-    #    hostssl p0x    all        0.0.0.0/0               cert    map=test
-    # where "test" map is defined in pgident.conf as
-    #    test            pgbouncer.acme.org      someuser
-    #    test            pgbouncer.acme.org      anotheruser
-    # hence the test succeeds.
-    bouncer.psql_test(
-        dbname="p0x",
-        host="localhost",
-        user="someuser",
-        sslmode="verify-full",
-        sslkey=client_key,
-        sslcert=client_cert,
-        sslrootcert=root,
-    )
-
-    bouncer.pg.sql("create user anotheruser with login;")
-
-    # The client connects to p0x using a client certificate with CN=pgbouncer.acme.org.
-    # hba_eval returns the followign line:
-    #    hostssl p0x    all        0.0.0.0/0               cert    map=test
-    # where "test" map is defined in pgident.conf as
-    #    test            pgbouncer.acme.org      someuser
-    #    test            pgbouncer.acme.org      anotheruser
-    # hence the test succeeds.
-    bouncer.psql_test(
-        dbname="p0x",
-        host="localhost",
-        user="anotheruser",
-        sslmode="verify-full",
-        sslkey=client_key,
-        sslcert=client_cert,
-        sslrootcert=root,
-    )
-
-    # The client connects to p0x using a client certificate with CN=pgbouncer.acme.org.
-    # hba_eval returns the followign line:
-    #    hostssl p0x    all        0.0.0.0/0               cert    map=test
-    # where "test" map is defined in pgident.conf as
-    #    test            pgbouncer.acme.org      someuser
-    #    test            pgbouncer.acme.org      anotheruser
-    # the username 'bouncer' does not match any mapped pg-username.
-    # hence the test fails.
-    with pytest.raises(
-        subprocess.CalledProcessError,
-    ):
-        with bouncer.log_contains(
-            "p0x/bouncer@127.0.0.1:43544 ident map: test does not have a match"
-        ):
-            bouncer.psql_test(
-                dbname="p0x",
-                host="localhost",
-                user="bouncer",
-                sslmode="verify-full",
-                sslkey=client_key,
-                sslcert=client_cert,
-                sslrootcert=root,
-            )
-
-    client_key = cert_dir / "TestCA1" / "sites" / "02-bouncer.key"
-    client_cert = cert_dir / "TestCA1" / "sites" / "02-bouncer.crt"
-
-    # The client connects to p0 using a client certificate with CN=bouncer.
-    # hba_eval returns the followign line:
-    #    hostssl p0              bouncer         0.0.0.0/0               cert
-    # CN expected in map is "bouncer" which matches the CN in the client cert
-    # hence the test succeeds.
-    bouncer.psql_test(
-        dbname="p0",
-        host="localhost",
-        user="bouncer",
-        sslmode="verify-full",
-        sslkey=client_key,
-        sslcert=client_cert,
-        sslrootcert=root,
-    )
-
-    # The client connects to p0y using a client certificate with CN=bouncer.
-    # hba_eval returns the followign line:
-    #    hostssl p0y             all             0.0.0.0/0               cert    map=test2
-    # where
-    #   test2           bouncer                 all
-    #   test2           pgbouncer.acme.org      "anotheruser"
-    # test2 mapping allows any client with CN="bouncer" to connect using any user name.
-    # Hence the test succeeds.
-    bouncer.psql_test(
-        dbname="p0y",
-        host="localhost",
-        user="someuser",
-        sslmode="verify-full",
-        sslkey=client_key,
-        sslcert=client_cert,
-        sslrootcert=root,
-    )
-
-    client_key = cert_dir / "TestCA1" / "sites" / "04-pgbouncer.acme.org.key"
-    client_cert = cert_dir / "TestCA1" / "sites" / "04-pgbouncer.acme.org.crt"
-
-    # The client connects to p0y using a client certificate with CN=pgbouncer.acme.org.
-    # hba_eval returns the followign line:
-    #    hostssl p0y             all             0.0.0.0/0               cert    map=test2
-    # where
-    #   test2           bouncer                 all
-    #   test2           pgbouncer.acme.org      "anotheruser"
-    # for CN=pgbouncer.acme.org, test2 allows to use anotheruser. Hence the test fails.
-
-    with pytest.raises(
-        subprocess.CalledProcessError,
-    ):
-        with bouncer.log_contains(
-            "p0y/someuser@127.0.0.1:39712 ident map: test2 does not have a match"
-        ):
-            bouncer.psql_test(
-                dbname="p0y",
-                host="localhost",
-                user="someuser",
-                sslmode="verify-full",
-                sslkey=client_key,
-                sslcert=client_cert,
-                sslrootcert=root,
-            )
-
-    # The client connects to p0y using a client certificate with CN=pgbouncer.acme.org.
-    # hba_eval returns the followign line:
-    #    hostssl p0y             all             0.0.0.0/0               cert    map=test2
-    # where
-    #   test2           bouncer                 all
-    #   test2           pgbouncer.acme.org      "anotheruser"
-    # for CN=pgbouncer.acme.org, test2 allows to use anotheruser. Hence the test succeeds.
-
-    bouncer.psql_test(
-        dbname="p0y",
-        host="localhost",
-        user="anotheruser",
-        sslmode="verify-full",
-        sslkey=client_key,
-        sslcert=client_cert,
-        sslrootcert=root,
-    )
-
-
-@pytest.mark.skipif("WINDOWS", reason="Windows does not have peer authentication")
-def test_peer_auth_ident_map(bouncer):
-    cur_user = getpass.getuser()
-
-    ident_conf_file = bouncer.config_dir / "ident.conf"
-    hba_conf_file = bouncer.config_dir / "hba.conf"
-
-    with open(ident_conf_file, "w") as f:
-        f.write(f"mymap {cur_user} postgres\n")
-        f.write(f"mymap {cur_user} someuser\n")
-
-    with open(hba_conf_file, "w") as f:
-        f.write(f"local   all  all peer map=mymap")
-
-    bouncer.write_ini(f"auth_type = hba")
-    bouncer.write_ini(
-        f"auth_query = SELECT usename, passwd FROM pg_shadow where usename = $1"
-    )
-    bouncer.write_ini(f"auth_user = pswcheck")
-    bouncer.write_ini(f"auth_file = {bouncer.auth_path}")
-    bouncer.write_ini(f"auth_hba_file = {hba_conf_file}")
-    bouncer.write_ini(f"auth_ident_file = {ident_conf_file}")
-
-    bouncer.admin("reload")
-
-    bouncer.psql_test(
-        dbname="p0y",
-        host=f"{bouncer.admin_host}",
-        user="postgres",
-    )
-
-    bouncer.psql_test(
-        dbname="p0y",
-        host=f"{bouncer.admin_host}",
-        user="someuser",
-    )
-
-    with pytest.raises(
-        subprocess.CalledProcessError,
-    ):
-        with bouncer.log_contains(
-            "p0y/bouncer@unix(6202):10202 ident map mymap cannot be matched"
-        ):
-            bouncer.psql_test(
-                dbname="p0y",
-                host=f"{bouncer.admin_host}",
-                user="bouncer",
-            )
-
-    with open(ident_conf_file, "w") as f:
-        f.write(f"mymap {cur_user} all")
-
-    bouncer.admin("reload")
-
-    bouncer.psql_test(
-        dbname="p0",
-        host=f"{bouncer.admin_host}",
-        user="bouncer",
-    )
